# Avoid using PeekBuffer for newline searching