PeekBuffer creates a lot of troubles when using ByteBuffer, but even with a byte array, switching to an indexOf-based algorithm significantly improves performance.